fix: return bool, remove async from new
