+++ node-uniffi/src/lib.rs
@@ -10,17 +10,13 @@ mod types;
 use celestia_types::ExtendedHeader;
 use error::{LuminaError, Result};
 use lumina_node::{
-    blockstore::RedbBlockstore,
-    events::{EventSubscriber, TryRecvError},
-    network::Network,
-    node::PeerTrackerInfo,
-    store::RedbStore,
+    blockstore::RedbBlockstore, events::EventSubscriber, node::PeerTrackerInfo, store::RedbStore,
     Node,
 };
-use std::{str::FromStr, sync::Arc};
+use std::str::FromStr;
 use tendermint::hash::Hash;
-use tokio::sync::Mutex;
-use types::{NetworkInfo, NodeEvent, NodeStartConfig, PeerId, SyncingInfo};
+use tokio::sync::{Mutex, RwLock};
+use types::{NetworkInfo, NodeConfig, NodeEvent, PeerId, SyncingInfo};
 use uniffi::Object;
 
 uniffi::setup_scaffolding!();
@@ -30,129 +26,134 @@ lumina_node::uniffi_reexport_scaffolding!();
 /// The main Lumina node that manages the connection to the Celestia network.
 #[derive(Object)]
 pub struct LuminaNode {
-    node: Arc<Mutex<Option<Node<RedbBlockstore, RedbStore>>>>,
-    network: Network,
-    events_subscriber: Arc<Mutex<Option<EventSubscriber>>>,
+    node: RwLock<Option<Node<RedbBlockstore, RedbStore>>>,
+    events_subscriber: Mutex<Option<EventSubscriber>>,
+    config: NodeConfig,
 }
 
 #[uniffi::export(async_runtime = "tokio")]
 impl LuminaNode {
     /// Sets a new connection to the Lumina node for the specified network.
     #[uniffi::constructor]
-    pub fn new(network: Network) -> Result<Arc<Self>> {
-        Ok(Arc::new(Self {
-            node: Arc::new(Mutex::new(None)),
-            network,
-            events_subscriber: Arc::new(Mutex::new(None)),
-        }))
+    pub fn new(config: NodeConfig) -> Result<Self> {
+        Ok(Self {
+            node: RwLock::new(None),
+            events_subscriber: Mutex::new(None),
+            config,
+        })
     }
 
-    /// Start the node without optional configuration.
-    /// UniFFI needs explicit handling for optional parameters to generate correct bindings for different languages.
+    /// Starts the node and connects to the network.
     pub async fn start(&self) -> Result<bool> {
-        self.start_with_config(None).await
-    }
-
-    /// Start the node with specific configuration
-    pub async fn start_with_config(&self, config: Option<NodeStartConfig>) -> Result<bool> {
-        let mut node_guard = self.node.lock().await;
-
-        if node_guard.is_some() {
-            return Err(LuminaError::AlreadyRunning);
+        let mut node_lock = self.node.write().await;
+        if node_lock.is_some() {
+            return Err(LuminaError::NetworkError {
+                msg: "Node is already running".to_string(),
+            });
         }
 
-        let config = config.unwrap_or_else(|| NodeStartConfig {
-            network: self.network.clone(),
-            bootnodes: None,
-            syncing_window_secs: None,
-            pruning_delay_secs: None,
-            batch_size: None,
-            ed25519_secret_key_bytes: None,
-        });
-
-        let builder = config.into_node_builder().await?;
-
+        let builder = self.config.clone().into_node_builder().await?;
         let (new_node, subscriber) = builder
             .start_subscribed()
             .await
             .map_err(|e| LuminaError::NetworkError { msg: e.to_string() })?;
 
-        let mut events_guard = self.events_subscriber.lock().await;
+        *self.events_subscriber.lock().await = Some(subscriber);
+        *node_lock = Some(new_node);
 
-        *events_guard = Some(subscriber);
-        *node_guard = Some(new_node);
         Ok(true)
     }
 
     /// Stops the running node and closes all network connections.
     pub async fn stop(&self) -> Result<()> {
-        let mut node_guard = self.node.lock().await;
-        let node = node_guard.take().ok_or(LuminaError::NodeNotRunning)?;
-        node.stop().await;
-        Ok(())
+        let mut node = self.node.write().await;
+        if let Some(node) = node.take() {
+            node.stop().await;
+            Ok(())
+        } else {
+            Err(LuminaError::NetworkError {
+                msg: "Node is already stopped".to_string(),
+            })
+        }
     }
 
     /// Checks if the node is currently running.
     pub async fn is_running(&self) -> bool {
-        self.node.lock().await.is_some()
+        self.node.read().await.is_some()
     }
 
     /// Gets the local peer ID as a string.
     pub async fn local_peer_id(&self) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         Ok(node.local_peer_id().to_base58())
     }
 
     /// Gets information about connected peers.
     pub async fn peer_tracker_info(&self) -> Result<PeerTrackerInfo> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         Ok(node.peer_tracker_info())
     }
 
     /// Waits until the node is connected to at least one peer.
     pub async fn wait_connected(&self) -> Result<()> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         Ok(node.wait_connected().await?)
     }
 
     /// Waits until the node is connected to at least one trusted peer.
     pub async fn wait_connected_trusted(&self) -> Result<()> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         Ok(node.wait_connected_trusted().await?)
     }
 
     /// Gets current network information.
     pub async fn network_info(&self) -> Result<NetworkInfo> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let info = node.network_info().await?;
         Ok(info.into())
     }
 
     /// Gets list of addresses the node is listening to.
     pub async fn listeners(&self) -> Result<Vec<String>> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let listeners = node.listeners().await?;
         Ok(listeners.into_iter().map(|l| l.to_string()).collect())
     }
 
     /// Gets list of currently connected peer IDs.
     pub async fn connected_peers(&self) -> Result<Vec<PeerId>> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let peers = node.connected_peers().await?;
         Ok(peers.into_iter().map(PeerId::from).collect())
     }
 
     /// Sets whether a peer with give ID is trusted.
     pub async fn set_peer_trust(&self, peer_id: PeerId, is_trusted: bool) -> Result<()> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let peer_id = peer_id
             .to_libp2p()
             .map_err(|e| LuminaError::NetworkError { msg: e.to_string() })?;
@@ -163,16 +164,20 @@ impl LuminaNode {
     ///
     /// Returns a serialized ExtendedHeader string.
     pub async fn request_head_header(&self) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let header = node.request_head_header().await?;
         Ok(header.to_string()) //if extended header is needed, we need a wrapper
     }
 
     /// Request a header for the block with a given hash from the network.
     pub async fn request_header_by_hash(&self, hash: String) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let hash =
             Hash::from_str(&hash).map_err(|e| LuminaError::InvalidHash { msg: e.to_string() })?;
         let header = node.request_header_by_hash(&hash).await?;
@@ -181,23 +186,27 @@ impl LuminaNode {
 
     /// Requests a header by its height.
     pub async fn request_header_by_height(&self, height: u64) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let header = node.request_header_by_height(height).await?;
         Ok(header.to_string())
     }
 
     /// Request headers in range (from, from + amount] from the network.
     ///
-    /// The headers will be verified with the `from` header.
+    /// The headers will be verified with the from header.
     /// Returns array of serialized ExtendedHeader strings.
     pub async fn request_verified_headers(
         &self,
         from: String, // serialized header like its done for WASM
         amount: u64,
     ) -> Result<Vec<String>> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let from: ExtendedHeader =
             serde_json::from_str(&from).map_err(|e| LuminaError::InvalidHeader {
                 msg: format!("Invalid header JSON: {}", e),
@@ -208,16 +217,20 @@ impl LuminaNode {
 
     /// Gets current syncing information.
     pub async fn syncer_info(&self) -> Result<SyncingInfo> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let info = node.syncer_info().await?;
         Ok(info.into())
     }
 
     /// Gets the latest header announced in the network.
     pub async fn get_network_head_header(&self) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let header = node.get_network_head_header().await?;
         header.map_or(
             // todo: better error handling, its undefined in wasm
@@ -230,16 +243,20 @@ impl LuminaNode {
 
     /// Gets the latest locally synced header.
     pub async fn get_local_head_header(&self) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let header = node.get_local_head_header().await?;
         Ok(header.to_string())
     }
 
     /// Get a synced header for the block with a given hash.
     pub async fn get_header_by_hash(&self, hash: String) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let hash =
             Hash::from_str(&hash).map_err(|e| LuminaError::InvalidHash { msg: e.to_string() })?;
         let header = node.get_header_by_hash(&hash).await?;
@@ -248,8 +265,10 @@ impl LuminaNode {
 
     /// Get a synced header for the block with a given height.
     pub async fn get_header_by_height(&self, height: u64) -> Result<String> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
         let header = node.get_header_by_height(height).await?;
         Ok(header.to_string())
     }
@@ -266,8 +285,10 @@ impl LuminaNode {
         start_height: Option<u64>,
         end_height: Option<u64>,
     ) -> Result<Vec<String>> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
 
         let headers = match (start_height, end_height) {
             (None, None) => node.get_headers(..).await,
@@ -283,21 +304,26 @@ impl LuminaNode {
     ///
     /// Returns serialized SamplingMetadata string if metadata exists for the height.
     pub async fn get_sampling_metadata(&self, height: u64) -> Result<Option<String>> {
-        let node_guard = self.node.lock().await;
-        let node = node_guard.as_ref().ok_or(LuminaError::NodeNotRunning)?;
+        let node = self.node.read().await;
+        let node = node.as_ref().ok_or(LuminaError::NetworkError {
+            msg: "Node not initialized".to_string(),
+        })?;
+
         let metadata = node.get_sampling_metadata(height).await?;
         Ok(metadata.map(|m| serde_json::to_string(&m).unwrap()))
     }
 
     /// Returns the next event from the node's event channel.
     pub async fn events_channel(&self) -> Result<Option<NodeEvent>> {
-        let mut events_guard = self.events_subscriber.lock().await;
-        let subscriber = events_guard.as_mut().ok_or(LuminaError::NodeNotRunning)?;
-
-        match subscriber.try_recv() {
-            Ok(event) => Ok(Some(event.event.into())),
-            Err(TryRecvError::Empty) => Ok(None),
-            Err(e) => Err(LuminaError::NetworkError { msg: e.to_string() }),
+        let mut events_subscriber = self.events_subscriber.lock().await;
+        match events_subscriber.as_mut() {
+            Some(subscriber) => match subscriber.try_recv() {
+                Ok(event) => Ok(Some(event.event.into())),
+                Err(e) => Err(LuminaError::NetworkError { msg: e.to_string() }),
+            },
+            None => Err(LuminaError::NetworkError {
+                msg: "Node is not running".to_string(),
+            }),
         }
     }
 }
