+++ node-uniffi/src/types/mod.rs
@@ -1,0 +1,9 @@
+mod config;
+mod event;
+mod network;
+mod sync;
+
+pub use config::NodeConfig;
+pub use event::{NodeEvent, PeerId};
+pub use network::NetworkInfo;
+pub use sync::SyncingInfo;
